Scanner can now add multiple series based on folder structure.
Set the getBookDataFromDir function to use all folder levels instead of just the last 3.  Top folder is author, bottom folder is book, and each folder between is added as a series in the series array.

Set the setBookMetadataFromFilenames function to return an array of series names instead of just one string.  In the array, the first series will get the sequence and the rest will get null as the sequence